fixing unit tests

+++ NextPvrWebConsole.Tests/NextPvrWebConsoleTest.cs
@@ -77,8 +77,11 @@ namespace NextPvrWebConsole.Tests
             string backupDb = System.IO.Path.Combine(npvrDir, "npvr.unittest_backup.db3");
             string nextPvrDbFile = System.IO.Path.Combine(npvrDir, settingsHelper.GetDatabaseFilename());
             // restore the db file
-            System.IO.File.Copy(backupDb, nextPvrDbFile, true);
-            System.IO.File.Delete(backupDb);
+            if (System.IO.File.Exists(nextPvrDbFile))
+            {
+                System.IO.File.Copy(backupDb, nextPvrDbFile, true);
+                System.IO.File.Delete(backupDb);
+            }
         }
 
         public virtual void Startup() { }
